Intermediate commit for conversion to a Windows service model.
Split the project into two binaries - one to run as a service and another to do the core functionality. This works toward eliminating the need for PsExec.
This commit will compile, but the install script does not reflect the changes and will fail.
If the service source is changed to specify a file path to the core binary, it will launch it as SYSTEM on the interactive desktop without the need for PsExec. There is currently no other way to actually specify the path to that binary.
Disabled the WIP code since it breaks things now.
Fixed a regression where handles to processes were not closed during the enumeration of windows.
Removed some old commented-out code.