Intermediate commit for elimination of PsExec.
Added proper logging for the change to purely file-based logging.
The project now has 3 crates with the addition of a shared one between the service and the core application.
The names of each crate have been changed to better reflect this separation.
The dependency on an external crate has been eliminated.
Changed the Rust edition of fsc_core to 2024 for consistency and resolved the lints.