Intermediate commit for the move to a service model.
The install script has been completely revamped to install the service instead of creating a scheduled task and all the code to get PsExec has been removed.
For the time being, the name of the service will have to be passed as a command line argument to the service. This is reflected in the install script.
The path to the core is now assumed to be 'fsc_core.exe' in the same directory as the service. This may change in the future.