First beta for 0.4.0. The core is now spawned correctly.
The service uses the token from winlogon to spawn the core with the correct setup to modify the SC/RM hider window.
Consolidated some error handling in the service.
The build script checks for version consistency again.
Fixed the install script to set the application path correctly.
Set up creation of the executable manifest for the separate service and core at build time.
Fixed up some logic in the core to enumerate windows properly.
Set up the common project to also statically include the C runtime.
Updated the README to reflect the current state of the project.
Set some explicit optimizations for release builds (matching NoBlockInput).